Genestealer conversion no longer instant
Changes genestealer conversion so it requires the genestealer to stay with the person (either on top or within 1 tile if they have their first evolution) for 10 seconds but is instant once this is done; could be increased to 15 or 20. Haven't been able to fix the runtime error yet but it's possible that it's coming up on my own tests beecause the victims don't have players in when converted.