Missile ammo and tomb entrance moved
Old/New
Krieg have 4 fake missiles and 1 launcher in their armoury/Krieg have 4 krak missiles and 1 launcher in their armoury
Garrison have no missiles and 1 launcher in their armoury/Garrison have 3 krak missiles and 1 launcher in their armoury. There is also a crate in the vault (that needs to be opened by the RT) that contains a single plasma missile.
Ultramarines have 1 krak missile and 6 launchers in their armoury/Ultramarines have 8 krak missiles and 4 launchers in their armoury

The hole to the tomb is now in a different part of the sewer.