Surgical Mechandrite recipe and examine fix
From what I can tell, the surgical mechandrite was added after the other mechandrites so I've added a recipe to it. Keeping with the same theme, I've used an exosuit module for the surgical mechandrite; this module is the syringe gun but the sleeper could also work.
The examine text for the incomplete autocannon mechandrite says to use an autocannon but it actually uses an AutoGun so I've corrected the text so it asks for the exosuit module it actually wants.