Added Anti-Tyranid and Anti-Orkoid chems
Anti-Orkoid chems do some damage on touch but most in the bloodstream. As they're not canon, the recipe has been commented out.
Anti-Tyranid do a lot of damage on touch but none in the bloodstream (as enabling them to consume reagents in their blood would break venomthropes). Evolved venomthropes could be highjacked if you inject them so they produce the anti-Tyranid chem. It does less damage to venomthropes but will kill them too.